Refactoring. Move transformRuntimeFunctionTypeToSuspendFunction to TypeDeserializer
